pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "massgen"
dynamic = ["version"]
description = "Multi-Agent Scaling System - A powerful framework for collaborative AI"
readme = { file = "README_PYPI.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "MassGen Team", email = "contact@massgen.dev" }
]
keywords = ["ai", "multi-agent", "collaboration", "orchestration", "llm", "claude", "gemini", "gpt", "xai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "datasets==3.2.0",
    "openai==2.2.0",
    "xai-sdk==1.0.0",
    "anthropic>=0.61.0",
    "cerebras-cloud-sdk==1.46.0",
    "lmstudio==1.4.1",
    "wcwidth>=0.2.5",
    "google-genai>=1.27.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "rich==14.1.0",
    "questionary>=2.0.0",
    "requests>=2.31.0",
    "typing-extensions>=4.0.0",
    "claude-agent-sdk>=0.1.21",
    "loguru>=0.7.0",
    "mcp>=1.12.0",
    "aiohttp>=3.8.0",
    "nest-asyncio>=1.6.0",
    "fastmcp>=2.12.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-book-theme>=1.1.4",
    "pillow>=11.3.0",
    "ag2>=0.9.10",
    "pyautogen>=0.10.0",
    "google-cloud-aiplatform>=1.71.1",
    "pytest>=8.4.2",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-core>=1.0.0",
    "agentscope>=1.0.6",
    "smolagents[litellm]>=1.22.0",
    "python-docx>=1.2.0",
    "openpyxl>=3.1.5",
    "python-pptx>=1.0.2",
    "opencv-python>=4.12.0.88",
    "pypdf2>=3.0.1",
    "mem0ai>=1.0.0",
    "reportlab>=4.0.0",
    "dspy>=2.4.0",
    "textual>=0.47.0",
    "litellm>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
observability = [
    "logfire>=3.0.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "autoflake>=2.1.0",
    "pyupgrade>=3.7.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-book-theme>=1.1.0",
    "myst-parser>=2.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-autobuild>=2021.3.14",
    "sphinxcontrib-mermaid>=0.9.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "sphinx-autoapi>=3.0.0",
    "sphinx-design>=0.5.0",
    "sphinx-tabs>=3.4.0",
    "sphinx-togglebutton>=0.3.0",
    "sphinx-inline-tabs>=2023.4.21",
    "sphinxcontrib-lunrsearch>=0.4",
    "sphinx-hoverxref>=1.3.0",
]
docker = [
    "docker>=7.0.0",
]
external = [
    "ag2>=0.9.10",
]
litellm = [
    "litellm>=1.0.0",
]
all = [
    "litellm>=1.0.0",
    "logfire>=3.0.0",
]

[project.scripts]
massgen = "massgen.cli:cli_main"

[project.urls]
Homepage = "https://github.com/Leezekun/MassGen"
Repository = "https://github.com/Leezekun/MassGen"
"Bug Reports" = "https://github.com/Leezekun/MassGen/issues"
Source = "https://github.com/Leezekun/MassGen"
Documentation = "https://docs.massgen.ai/en/latest/"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.dynamic]
version = {attr = "massgen.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "massgen*"]
exclude = [
    "tests*",
    "docs*",
    "future_mass*",
    "massgen_logs*",
    "snapshots*",
    "temp_workspaces*",
    "workspace*",
    "sessions*",
    ".massgen*",
    "claude_code_workspace*",
    "code_execution_workspaces*",
    "agent_outputs*",
    "traces*",
    "summaries*",
    "output*",
    ".playwright-mcp*",
]

[tool.setuptools.package-data]
massgen = [
    "configs/**/*.yaml",
    "configs/**/*.yml",
    "examples/*.yaml",
    "backends/.env.example",
    "tool/**/*.py",
    "tool/**/*.yaml",
    "tool/**/*.md",
    "skills/**/*",
    "frontend/web/static/**/*",
]
"*" = ["*.json", "*.yaml", "*.yml", "*.md"]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | future_mass
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["future_mass/*"]

# flake8 configuration (in setup.cfg since flake8 doesn't support pyproject.toml yet)
# See .flake8 file for configuration

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "agents.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "future_mass.*"
ignore_errors = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["massgen/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: mark test as requiring asyncio",
    "docker: marks tests that require Docker to be installed and running",
    "expensive: marks tests that cost money (API calls with >$0.01 cost, skip by default)",
    "performance: marks tests as performance tests (run manually when needed)",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pytest.PytestAssertRewriteWarning:_pytest.config",
]

# Coverage configuration
[tool.coverage.run]
source = ["agents", "massgen_*"]
omit = [
    "*/tests/*",
    "*/test_*",
    "future_mass/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", "future_mass"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection for test files

[dependency-groups]
dev = [
    "astroid>=3.0.0",
    "pre-commit>=4.3.0",
    "pylint>=3.0.0",
    "linkify-it-py>=2.0.3",
    "myst-parser>=4.0.1",
    "sphinx>=8.1.3",
    "sphinx-copybutton>=0.5.2",
    "sphinx-design>=0.6.1",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-book-theme>=1.1.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
